fix - search page refreshing

+++ src/sections/search/UserItem.tsx
@@ -6,6 +6,8 @@ import { getNameInitials } from "@util/string";
 import http from "../../client/axios";
 import useAuth from "@hook/useAuth";
 import { useRouter } from "next/router";
+import { useModalAction } from "../../contexts/ModalContext";
+import { BsPersonPlus } from "react-icons/bs";
 
 interface Props {
   id?: string;
@@ -16,12 +18,21 @@ interface Props {
 }
 
 const SearchUserItem = ({ id, name, username, profile, following }: Props) => {
-  const { user } = useAuth();
+  const { user, isAuthenticated } = useAuth();
   const [isFollowing, setIsFollowing] = useState(following);
   const [hover, setHover] = useState(false);
   const router = useRouter();
+  const { openModal } = useModalAction();
   // un/follow
   async function handleFollow() {
+    if (!isAuthenticated) {
+      openModal("AUTH_BOARDING", {
+        icon: <BsPersonPlus className="w-12 h-12 text-[#1da1f2]" />,
+        title: `Follow ${name} to see what they share on Twitter.`,
+        desc: "Sign up so you never miss their Tweets.",
+      });
+      return;
+    }
     setIsFollowing(!isFollowing);
     if (isFollowing) await http.delete(`/follow/${id}`);
     else await http.post(`/follow/${id}`);
